backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-crispy-forms==2.3
environs==11.0.0
marshmallow==3.22.0
packaging==24.1
pillow==10.3.0  # вместо 11.0.0
python-dotenv==1.0.1
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
wheel==0.44.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0